backend/requirements.txt
annotated-types==0.7.0
anyio>=4.8.0,<5.0.0
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.116.1
h11==0.14.0
idna==3.4
pydantic==2.11.7
pydantic_core==2.33.2
requests>=2.28.1,<3.0.0
sniffio==1.3.0
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==1.26.15
uvicorn==0.21.1
python-dotenv
google-genai
psycopg2-binary